Bugfix- explode fail after overriding the archive filename
